donate-crypto/pyproject.toml
[tool.poetry]
name = "donate-crypto"
version = "0.1.0"
description = "Crypto Donations"
authors = ["Fedor Saveliev <nett_ef@icloud.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.12"


[tool.poetry.group.dev.dependencies]
pyyaml = "^6.0.1"
jinja2 = "^3.1.3"
qrcode = "^7.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

